CSN: CSN-hasim-mips-128
The functional partition had a bug in regstate/freelist, where it rewinds to a token before servicing all existing requests that exist before rewind. This made it go to 
a dirty state, leading to errors. I fixed that here. So it passed all the tests after that. But then, Michael found a bug in the functional partition's store buffer
where it didnt get initialized in the beginning. So he added an initialization state, but this broke something else in the store buffer, so it doesnt pass the store test now.
I am currently debugging it. All the tests pass in svn version 203