CSN: CSN-hasim-mips-132
Checked if all branches work. Came across this issue where the functional partition doesn't differentiate between a not taken branch and a sequential instruction. So had to do some workaround for that ( carrying the current pc around)
Also looks like the functional parittion has a bug with respect to finishing program. I kept getting finished programs from FP, even for non MTC0 instructions. So I made a change where I check for finished program only in MTC0 instructions. So now all branch predictors pass. ( Didnt check global history predictor)

I think one more change is needed - changing the queues to use BRAM. Also the latency can be effectively hidden even in the case of BRAMs because the reads happen in order, so we can prefetch data in the head of the bram queue. I am writing a data structure which captures this. I will be checking that in shortly