CSN: CSN-hasim-mips-140
Added a reorder buffer in the decode stage and the execute stage. The FuncP is not in-order at these stages ( and i was simply luck to have not been hit by this before. I caught this when I changed the schedule a bit, making commit go more in parallel ). Now commit goes in parallel with the decode, which is not much because commit was never in the critical path, but it sure helped me catch a bug where I assumed FuncP is completely in-order