CSN: CSN-hasim-mips-147
Fully integrated the s10k model with the new functional partion. Michael found a bug in the execute stage of the functional partion and he fixed it temporarily in this code. So now this passes all benchmarks except qsort and towers. Michael identified the problem with the store buffer which lead to infinite traversals...