CSN: CSN-hasim-mips-176
Added a preliminary version of V3 of the SMIPS ISA. This version passes all ISA
tests and 2/5 long-running benchmarks, so is not quite ready for prime time yet.

Updated more modules to the new coding conventions.

Also added a stopgap measure to the benchmark scripts until the benchmark system
is overhauled.


Michael P.